feat: refactor: move follow related functions to follow-controller
