+++ controllers/user-controller.js
@@ -3,7 +3,7 @@ const jwt = require('jsonwebtoken')
 const sequelize = require('sequelize')
 
 const helpers = require('../_helpers')
-const { User, Followship, Tweet, Like, Reply } = require('../models')
+const { User, Tweet, Like, Reply } = require('../models')
 
 const userController = {
   signin: async (req, res, next) => {
@@ -231,82 +231,6 @@ const userController = {
       }))
 
       res.status(200).json(tweetsIncludeIsLike)
-    } catch (err) {
-      next(err)
-    }
-  },
-  addFollow: async (req, res, next) => {
-    try {
-      const followerId = Number(helpers.getUser(req).id)
-      const followingId = Number(req.body.id)
-      if (!followerId || !followingId) {
-        return res.status(400).json({
-          status: 'error',
-          message: 'followerId and followingId required'
-        })
-      }
-
-      if (followerId === followingId) {
-        return res.status(401).json({
-          status: 'error',
-          message: 'Can not follow yourself.'
-        })
-      }
-
-      const followed = await Followship.findOne({
-        where: { followerId, followingId }
-      })
-      if (followed) {
-        return res.status(401).json({
-          status: 'error',
-          message: 'Already followed'
-        })
-      }
-
-      await Followship.create({ followerId, followingId })
-      return res.status(200).json({
-        status: 'success',
-        message: 'Followship added'
-      })
-    } catch (err) {
-      next(err)
-    }
-  },
-  removeFollow: async (req, res, next) => {
-    try {
-      const followerId = Number(helpers.getUser(req).id)
-      const followingId = Number(req.params.followingId)
-      if (!followerId || !followingId) {
-        return res.status(400).json({
-          status: 'error',
-          message: 'followerId and followingId required'
-        })
-      }
-
-      const follower = await User.findByPk(followerId)
-      const following = await User.findByPk(followingId)
-      if (!follower || !following) {
-        return res.status(401).json({
-          status: 'error',
-          message: 'Follower or following not exists.'
-        })
-      }
-
-      const followship = await Followship.findOne({
-        where: { followerId, followingId }
-      })
-      if (!followship) {
-        return res.status(401).json({
-          status: 'error',
-          message: 'Not followed yet'
-        })
-      }
-
-      await followship.destroy()
-      return res.status(200).json({
-        status: 'success',
-        message: 'Remove followed success'
-      })
     } catch (err) {
       next(err)
     }
