+++ routes/modules/followships.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const router = express.Router()
 
-const userController = require('../../controllers/user-controller')
+const userController = require('../../controllers/follow-controller')
 
 router.delete('/:followingId', userController.removeFollow)
 router.post('/', userController.addFollow)
